Enhance post management and add new models
Updated Post class with new properties and methods for media file handling, publishing, and content processing. Simplified PostMember class by removing CreatedAt property. Changed id parameter type to long in IPostRepository and PostRepository. Refactored UserProfileApplication to use model conversion method. Added IPostApplication interface and PostApplication implementation for managing post operations. Introduced helper classes and models for error handling, event representation, and data mapping. Updated project file to include HtmlSanitizer dependency.

+++ src/HolaViaje.Social/Features/Posts/Repository/IPostRepository.cs
@@ -2,7 +2,7 @@
 
 public interface IPostRepository
 {
-    Task<Post?> GetAsync(int id, bool traking = false, CancellationToken cancellationToken = default);
+    Task<Post?> GetAsync(long id, bool traking = false, CancellationToken cancellationToken = default);
     Task<Post> CreateAsync(Post post, CancellationToken cancellationToken = default);
     Task<Post> UpdateAsync(Post post, CancellationToken cancellationToken = default);
 }
